refactor: Refatora o Controller do Ninja para incluir o responseEntiry nas minhas rotas e customizar a resposta das requisicoes do cliente

+++ src/main/java/dev/Java10x/CadastroDeNinjas/Ninjas/NinjaController.java
@@ -1,5 +1,8 @@
 package dev.Java10x.CadastroDeNinjas.Ninjas;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -21,32 +24,53 @@ public class NinjaController {
 
     // Adicionar ninja (CREATE)
     @PostMapping("/create")
-    public NinjaDTO createNinja(@RequestBody NinjaDTO ninja) {
-        return ninjaService.createNinja(ninja);
+    public ResponseEntity<String> createNinja(@RequestBody NinjaDTO ninja) {
+        NinjaDTO newNinja = ninjaService.createNinja(ninja);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Ninja criado com sucesso: " + newNinja.getName() + " (ID): " + newNinja.getId());
     }
 
     // Mostrar todos os ninjas (READ)
     @GetMapping("/list")
-    public List<NinjaDTO> listAllNinjas() {
-        return ninjaService.listAllNinjas();
+    public ResponseEntity<List<NinjaDTO>> listAllNinjas() {
+        List<NinjaDTO> ninjas = ninjaService.listAllNinjas();
+        return ResponseEntity.ok(ninjas);
     }
 
     // Mostrar ninja por ID (READ)
     @GetMapping("/list/{id}")
-    public NinjaDTO listNinjaById(@PathVariable Long id) {
-        return ninjaService.listNinjaById(id);
+    public ResponseEntity<?> listNinjaById(@PathVariable Long id) {
+        NinjaDTO ninja = ninjaService.listNinjaById(id);
+        if (ninja != null) {
+            return ResponseEntity.ok(ninja);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Ninja com o ID: " + id + " nao encontrado.");
+        }
     }
 
     // Alterar dados dos ninjas (UPDATE)
     @PutMapping("/update/{id}")
-    public NinjaDTO updateNinjaById(@PathVariable Long id, @RequestBody NinjaDTO updatedNinja) {
-        return ninjaService.updateNinja(id, updatedNinja);
+    public ResponseEntity<?> updateNinjaById(@PathVariable Long id, @RequestBody NinjaDTO updatedNinja) {
+        NinjaDTO ninja = ninjaService.updateNinja(id, updatedNinja);
+        if (ninja != null) {
+            return ResponseEntity.ok(ninja);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Ninja com o ID: " + id + " nao encontrado.");
+        }
     }
 
     // Deletar ninja (DELETE)
     @DeleteMapping("/delete/{id}")
-    public void deleteNinjaById(@PathVariable Long id) {
-        ninjaService.deleteNinjaById(id);
+    public ResponseEntity<String> deleteNinjaById(@PathVariable Long id) {
+        if (ninjaService.listNinjaById(id) != null) {
+            ninjaService.deleteNinjaById(id);
+            return ResponseEntity.ok("Ninja com o ID: " + id + " deletado com sucesso.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Ninja com o ID: " + id + " nao encontrado.");
+        }
     }
 
 }
